async: added mechanisms to re-open event file descriptors
After a fork, the child shares fds with its parent, which is
a problem because an event notification will affect unrelated
QEMU instances.

This commit adds support for re-creating these fds,
while ensuring that the new fd is the same as before. Having
the same fd is important because the fd may be used in other
data structures.

It is not possible to just re-create all async structures from
scratch because all the pointers would be different and we don't
really know where they are used in order to patch everything
properly.

Signed-off-by: Vitaly Chipounov <vitaly@cyberhaven.io>